pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Leonardo Melo <112220689+euleozera@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
factory-boy = "^3.2.1"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
